refactor: reward calculator
