+++ Cargo.toml
@@ -30,6 +30,7 @@ members = [
     "util/stop-handler",
     "util/app-config",
     "script/data-loader",
+    "util/reward-calculator",
     "shared/tx-pool-executor",
     "ckb-bin",
     "traits",
