+++ shared/Cargo.toml
@@ -26,6 +26,7 @@ ckb-protocol = { path = "../protocol" }
 ckb-script = { path = "../script" }
 occupied-capacity = { path = "../util/occupied-capacity" }
 dao-utils = { path = "../util/dao/utils" }
+reward-calculator = { path = "../util/reward-calculator" }
 
 [dev-dependencies]
 env_logger = "0.6"
